requirements.txt
flask==2.0.1
werkzeug==2.0.1
transformers==4.30.2
torch==2.0.1
opencv-python==4.7.0.72
numpy==1.24.3
google-generativeai==0.1.0